feat: Implement Realtime Manager for event synchronization and UI updates
- Added RealtimeManager class to handle real-time events between backend and frontend.
- Integrated event subscriptions for crawling progress, task status, stage changes, errors, database updates, and completion events.
- Implemented connection management with retry logic and event statistics tracking.
- Created crawlerStore for managing crawling-related state and actions.
- Developed databaseStore for handling database state and operations, including backup and optimization.
- Refactored uiStore to manage UI state and notifications separately from crawling logic.
- Enhanced error handling and user notifications throughout the application.

+++ src-tauri/src/domain/integrated_product.rs
@@ -197,7 +197,7 @@ impl IntegratedProduct {
         if self.size.is_some() { score += 2.0; }
         if self.material.is_some() { score += 2.0; }
 
-        ((score / max_score * 100.0) as f64).round()
+        (score / max_score * 100.0_f64).round()
     }
 
     /// 제품 정보 업데이트
